Refactor library. Remove initialization side-effects
