+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "axios-jwt",
-  "version": "2.1.0",
+  "version": "2.2.0",
   "description": "Axios interceptor to store, use, and refresh tokens for authentication.",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
