+++ src/IAuthTokenInterceptorConfig.ts
@@ -1,0 +1,8 @@
+import { TokenRefreshRequest } from './TokenRefreshRequest';
+
+export interface IAuthTokenInterceptorConfig {
+  header?: string;
+  headerPrefix?: string;
+  requestRefresh: TokenRefreshRequest;
+  tokenExpireFudge?: number
+}
