+++ src/IAuthTokens.ts
@@ -1,0 +1,6 @@
+import { Token } from './Token';
+
+export interface IAuthTokens {
+  accessToken: Token;
+  refreshToken: Token;
+}
