+++ src/StorageKey.ts
@@ -1,0 +1,1 @@
+export const STORAGE_KEY = `auth-tokens-${process.env.NODE_ENV}`;
