+++ src/Token.ts
@@ -1,0 +1,1 @@
+export type Token = string
