+++ src/TokenRefreshRequest.ts
@@ -1,0 +1,4 @@
+import { Token } from './Token';
+import { IAuthTokens } from './IAuthTokens';
+
+export type TokenRefreshRequest = (refreshToken: Token) => Promise<Token | IAuthTokens>
