+++ src/applyLocalStorage.ts
@@ -1,0 +1,8 @@
+import { WebStorageService } from './WebStorageService';
+import { applyStorage } from './applyStorage';
+
+export const applyLocalStorage = () => {
+  applyStorage(
+    new WebStorageService(window.localStorage)
+  )
+}
