+++ src/applySessionStorage.ts
@@ -1,0 +1,8 @@
+import { WebStorageService } from './WebStorageService';
+import { applyStorage } from './applyStorage';
+
+export const applySessionStorage = () => {
+  applyStorage(
+    new WebStorageService(window.sessionStorage)
+  )
+}
