+++ src/applyStorage.ts
@@ -1,0 +1,6 @@
+import { StorageProxy } from './StorageProxy';
+import { StorageServiceType } from './StorageServiceType';
+
+export const applyStorage = (storage: StorageServiceType) => {
+  StorageProxy.Storage = storage;
+};
