--- src/storage.ts
@@ -1,5 +1,0 @@
-import { StorageProxy } from './StorageProxy';
-import { WebStorageService } from './WebStorageService';
-
-// Initialize default web-storage
-StorageProxy.Storage = new WebStorageService(localStorage)
