+++ tests/clearAuthTokens.test.ts
@@ -1,6 +1,12 @@
-import { STORAGE_KEY, clearAuthTokens } from '../src'
+import { STORAGE_KEY } from '../src/StorageKey';
+import { applyLocalStorage, clearAuthTokens } from '../src';
+
 
 describe('clearAuthTokens', () => {
+  beforeAll(() => {
+    applyLocalStorage()
+  })
+
   it('removes the tokens from localstorage', () => {
     // GIVEN
     // Tokens are stored in localStorage
