+++ tests/refreshTokenIfNeeded.test.ts
@@ -1,6 +1,7 @@
-import { STORAGE_KEY, refreshTokenIfNeeded } from '../src'
-import jwt from 'jsonwebtoken'
+import { applyLocalStorage, refreshTokenIfNeeded } from '../src';
+import jwt from 'jsonwebtoken';
 import { AxiosError } from 'axios'
+import { STORAGE_KEY } from '../src/StorageKey';
 
 function makeAxiosErrorWithStatusCode(statusCode: number) {
   const error = new AxiosError(
@@ -21,6 +22,10 @@ function makeAxiosErrorWithStatusCode(statusCode: number) {
 }
 
 describe('refreshTokenIfNeeded', () => {
+  beforeAll(() => {
+    applyLocalStorage()
+  })
+
   it('throws an error if the requestRefresh function threw one', async () => {
     // GIVEN
     // I have an access token that expired an hour ago
