+++ tests/setAccessToken.test.ts
@@ -1,6 +1,11 @@
-import { STORAGE_KEY, setAccessToken } from '../src'
+import { applyLocalStorage, setAccessToken } from '../src';
+import { STORAGE_KEY } from '../src/StorageKey';
 
 describe('setAccessToken', () => {
+  beforeAll(() => {
+    applyLocalStorage()
+  })
+
   it('throws an error if there are no tokens stored', () => {
     // GIVEN
     // localStorage is empty
