+++ tests/setAuthTokens.test.ts
@@ -1,6 +1,11 @@
-import { STORAGE_KEY, setAuthTokens } from '../src'
+import { applyLocalStorage, setAuthTokens } from '../src';
+import { STORAGE_KEY } from '../src/StorageKey';
 
 describe('setAuthTokens', () => {
+  beforeAll(() => {
+    applyLocalStorage()
+  })
+
   it('stores the tokens in localstorage', () => {
     // GIVEN
     // localStorage is empty
